python-google-adk/pyproject.toml
[project]
name = "python-google-adk"
version = "0.0.1"
description = "Demo LLM Application with Google Agent Development Kit"
authors = [{ name = "nazna", email = "naznagg@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["duckdb>=1.3.2", "google-adk>=1.13.0", "pydantic>=2.11.7"]

[dependency-groups]
dev = ["google-genai>=1.32.0", "pandas>=2.3.2"]

[tool.uv]
python-downloads = "manual"

[tool.ruff]
fix = true
line-length = 120

[tool.ruff.lint.isort]
force-single-line = true
